dm vdo user: combine uds_init{,ialize}_mutex()
Historically, uds_initialize_mutex() was separate in order to allow a
second parameter, assert_on_error. This parameter was only checked in
combination with examining the return value of pthread_mutex_init().

However, to quote man pthread_mutex_init(3) on linux [1]:

RETURN VALUE

       pthread_mutex_init always returns 0.  The other mutex functions
       return 0 on success and a non-zero error code on error.

Hence, uds_initialize_mutex's second parameter never matters, and it can
be combined with uds_init_mutex, a thin wrapper around it.

There is one usage which wasn't checking the return value --
permassert.c -- but this actually was unnecessary at all because the
static initializer was already in use to initialize the mutex.

[1] https://git.kernel.org/pub/scm/docs/man-pages/man-pages.git/tree/man/man3/pthread_mutex_init.3

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>