uds: eliminate usage of UDS_MUTEX_KIND
This environment variable used to control what kind of mutex were in
use, in theory. It could be set to error-checking to use pthread's error
checking mutexes, or fast-adaptive to not do so. Or, the default was to
use pthread's fast adaptive mutexes on NDEBUG builds, and error-checking
on !NDEBUG builds.

However, there was inconsistency -- some mutexes are statically
initialized, and those would use the compile flag NDEBUG only. And, it's
particularly unclear, given the quickness of the build process, why it's
important to have a runtime variable for this particular debug option.

Therefore, this change makes the mutex kind entirely controlled by the
NDEBUG flag, eliminating the use of the environment variable.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>